Refactor the main App.jsx file into smaller, more manageable components to improve code organization and maintainability.
- Created a new directory structure under `src/` for components, constants, api, firebase, and icons.
- Broke down the single `App.jsx` file into multiple component files, each with a single responsibility.
- Moved all constants, such as color themes and grade data, into a new `src/constants` directory.
- Extracted Firebase configuration and helper functions into a dedicated `src/firebase` directory.
- Separated the Gemini API call into its own `src/api` module.
- The main `App.jsx` now serves as the primary layout component, importing all the child components.
