+++ src/constants/theme.js
@@ -1,0 +1,28 @@
+export const colorThemes = {
+    teal: { bg: 'bg-teal-500/10', border: 'border-teal-500/30', shadow: 'hover:shadow-teal-500/20', text: 'text-teal-300', hex: '#2dd4bf' },
+    sky: { bg: 'bg-sky-500/10', border: 'border-sky-500/30', shadow: 'hover:shadow-sky-500/20', text: 'text-sky-300', hex: '#38bdf8' },
+    purple: { bg: 'bg-purple-500/10', border: 'border-purple-500/30', shadow: 'hover:shadow-purple-500/20', text: 'text-purple-300', hex: '#a855f7' },
+    rose: { bg: 'bg-rose-500/10', border: 'border-rose-500/30', shadow: 'hover:shadow-rose-500/20', text: 'text-rose-300', hex: '#f43f5e' },
+    amber: { bg: 'bg-amber-500/10', border: 'border-amber-500/30', shadow: 'hover:shadow-amber-500/20', text: 'text-amber-300', hex: '#f59e0b' },
+    lime: { bg: 'bg-lime-500/10', border: 'border-lime-500/30', shadow: 'hover:shadow-lime-500/20', text: 'text-lime-300', hex: '#84cc16' },
+};
+
+export const gradeStyles = {
+    p1: { icon: 'Square', color: 'text-rose-400' },
+    p2: { icon: 'SquareDot', color: 'text-orange-400' },
+    p3: { icon: 'SquarePen', color: 'text-amber-400' },
+    p4: { icon: 'SquareAsterisk', color: 'text-lime-400' },
+    p5: { icon: 'SquareEqual', color: 'text-sky-400' },
+    p6: { icon: 'SquareM', color: 'text-purple-400' },
+};
+
+export const analyticsCardStyles = [
+    { gradient: 'from-purple-500/70 to-indigo-500/70', border: 'border-purple-400' },
+    { gradient: 'from-green-500/70 to-teal-500/70', border: 'border-green-400' },
+    { gradient: 'from-pink-500/70 to-rose-500/70', border: 'border-pink-400' },
+    { gradient: 'from-orange-500/70 to-amber-500/70', border: 'border-orange-400' },
+    { gradient: 'from-sky-500/70 to-cyan-500/70', border: 'border-sky-400' },
+    { gradient: 'from-red-500/70 to-orange-500/70', border: 'border-red-400' },
+    { gradient: 'from-yellow-500/70 to-lime-500/70', border: 'border-yellow-400' },
+    { gradient: 'from-fuchsia-500/70 to-purple-500/70', border: 'border-fuchsia-400' },
+];
